+++ src/python.cpp
@@ -3033,7 +3033,9 @@ SEXP py_list_length(PyObjectRef x) {
 SEXP py_len_impl(PyObjectRef x, SEXP defaultValue) {
 
   Py_ssize_t value = PyObject_Size(x);
-  if (PyErr_Occurred()) {
+  if (value == -1) {
+   // object is missing a `__len__` method, or a `__len__` method that
+   // intentionally raises an Exception
     if (defaultValue == R_NilValue) {
       stop(py_fetch_error());
     } else {
@@ -3046,20 +3048,19 @@ SEXP py_len_impl(PyObjectRef x, SEXP defaultValue) {
     return Rf_ScalarInteger((int) value);
   else
     return Rf_ScalarReal((double) value);
-
 }
 
 // [[Rcpp::export]]
-SEXP py_bool(PyObjectRef x) {
+SEXP py_bool_impl(PyObjectRef x) {
 
-  // invoke __bool__ method
-  PyObjectPtr result(PyObject_CallMethod(x, "__bool__", NULL));
-  if (PyErr_Occurred()) {
-    PyErr_Clear();
-    return Rf_ScalarLogical(0);
+  // evaluate Python `not not x`
+  int result = PyObject_IsTrue(x);
+
+  if (result == -1) {
+  // Should only happen if the object has a `__bool__` method that
+  // intentionally throws an exception.
+    stop(py_fetch_error());
   }
 
-  // check whether it's the True value
-  return Rf_ScalarLogical(result == Py_True);
-
+  return Rf_ScalarLogical(result);
 }